fixing rendering issue
