+++ src/views/404.html
@@ -1,0 +1,1 @@
+<h1>404</h1>
