+++ src/views/barbers.html
