+++ src/views/landing.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>BARBEXA - Premium Barbershop</title>
     <link rel="icon" type="image/png" href="/app/assets/img/logo2.png" />
+    <link rel="stylesheet" href="../css/landing.css">
 
     <!-- FontAwesome -->
     <link
@@ -234,6 +235,6 @@
     <!-- FullCalendar -->
     <script src="https://cdn.jsdelivr.net/npm/fullcalendar@6.1.8/main.min.js"></script>
     <!-- Main Script -->
-    <script type="module" src="/app/controllers/landing.js"></script>
+    
   </body>
 </html>
